Converts systems package to module Renames package to _systems, adds module, removes redundant init, and fixes imports
