+++ src/pyrite/_systems/base_system.py
