+++ src/pyrite/systems.py
@@ -1,0 +1,3 @@
+import pyrite._systems.base_system
+
+System = pyrite._systems.base_system.BaseSystem
