--- src/pyrite/systems/__init__.py
@@ -1,1 +1,0 @@
-from .base_system import BaseSystem as System  # noqa:F401
